update other usages
