+++ src/pages/FirstUsePage.js
@@ -23,6 +23,7 @@ import { importData } from '../database/utilities';
 
 import globalStyles, { SUSSOL_ORANGE, WARM_GREY } from '../globalStyles';
 import { FormPasswordInput } from '../widgets/FormInputs/FormPasswordInput';
+import { AuthFormView } from '../widgets/AuthFormView';
 
 const STATUSES = {
   UNINITIALISED: 'uninitialised',
@@ -118,7 +119,7 @@ export class FirstUsePageComponent extends React.Component {
 
     return (
       <View style={[globalStyles.verticalContainer, localStyles.verticalContainer]}>
-        <View style={globalStyles.authFormContainer}>
+        <AuthFormView>
           <Image
             resizeMode="contain"
             style={globalStyles.authFormLogo}
@@ -196,7 +197,7 @@ export class FirstUsePageComponent extends React.Component {
               isDisabled={!this.canAttemptLogin}
             />
           </View>
-        </View>
+        </AuthFormView>
         <View style={localStyles.demoSiteRequestButtonContainer}>
           <View style={globalStyles.horizontalContainer}>
             <Button
